fix statistics table
